Some refactoring for the base json class
