+++ stackify/__init__.py
@@ -9,12 +9,24 @@ API_URL = 'https://api.stackify.com'
 
 READ_TIMEOUT = 5000
 
+MAX_BATCH = 100
+
 
 import logging
+
+LOGGING_LEVELS = {
+    logging.CRITICAL: 'CRITICAL',
+    logging.ERROR: 'ERROR',
+    logging.WARNING: 'WARNING',
+    logging.INFO: 'INFO',
+    logging.DEBUG: 'DEBUG',
+    logging.NOTSET: 'NOTSET'
+}
 
 logging.basicConfig()
 
 internal_log = logging.getLogger(__name__)
+
 
 from stackify.application import ApiConfiguration
 from stackify.http import HTTPClient
