+++ stackify/formats.py
@@ -1,0 +1,11 @@
+import json
+
+
+def no_nones(d):
+    return {k: v for k,v in d.items() if v is not None}
+
+
+class JSONObject(object):
+    def toJSON(self):
+        return json.dumps(self, default=lambda x: no_nones(x.__dict__))
+
